+++ src/main/resources/shader/image_frag.glsl
@@ -1,0 +1,12 @@
+#version 330
+
+out vec4 color;
+
+in vec2 textire_coordinates;
+
+uniform sampler2D ourTexture;
+
+void main() {
+
+    color = texture(ourTexture, textire_coordinates);
+}
